refactor(template): rename rx-let.directive.ts to let.directive.ts
