+++ libs/template/let/src/lib/tests/let.directive.error.spec.ts
@@ -9,7 +9,7 @@ import { RX_RENDER_STRATEGIES_CONFIG } from '@rx-angular/cdk/render-strategies';
 import { mockConsole } from '@test-helpers';
 import { Observable, of, throwError } from 'rxjs';
 
-import { RxLet } from '../rx-let.directive';
+import { RxLet } from '../let.directive';
 import { MockChangeDetectorRef } from './fixtures';
 
 @Component({
