+++ libs/template/let/src/lib/tests/let.directive.next.spec.ts
@@ -7,7 +7,7 @@ import {
 import { RX_RENDER_STRATEGIES_CONFIG } from '@rx-angular/cdk/render-strategies';
 import { EMPTY, interval, NEVER, Observable, of } from 'rxjs';
 import { fakeAsync, TestBed, tick } from '@angular/core/testing';
-import { RxLet } from '../rx-let.directive';
+import { RxLet } from '../let.directive';
 import { take } from 'rxjs/operators';
 import { MockChangeDetectorRef } from './fixtures';
 import { mockConsole } from '@test-helpers';
